Create Demo_Touch_Keyboard
I needed a simple keyboard to use the GUI on a touchscreen system without keyboard. Turned out to be quite tricky and I do not think that this is the best way to do it. But in principle it works. I didn't manage to get it working with Multilines and Comboboxes though.

Maybe someone with better understanding of PySimpleGUI and tkinter can offer a more universal and elegant way...