Update PySimpleGUIQt.py
Updated functionality allowing the user to just specify 'underline' in the font argument if they want to keep the default font and font size.  As it stands, if you change the font argument to ("", "" ,"underline") or "   underline" it will return Arial font instead of the default Helvetica used by the rest of the default fonts.